Text Summarization/api/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
transformers==4.42.3
torch==2.0.1
protobuf<=3.20.3
